+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Naming things: Renamed argument `snap_days` to `snap_hours`
 - Formatting: Added formatter for Opsgenie API query clause
 - Formatting: Added formatter for GitHub Search API query clause
+- `TimeInterval`: Made `end` argument optional
 
 ## 2025-02-24 v0.2.1
 - Emit better error message when German locales are not installed